+++ src/bot_functions/fetching/finance/StockMarketFunctions.ts
